+++ __main__.py
@@ -4,23 +4,23 @@
 
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 # IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
 # AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
 
-"""PDFCrawler is a simple search application built using Python and PyQt5."""
+""" PDFCrawler is a simple search application built using Python and PyQt5. """
 
 import sys
 
-from model import *
-from view import *
-from controller import *
+from src.model import *
+from src.view import *
+from src.controller import *
 
-__version__ = "1.0"
-__author__ = "Klaus Moser"
+__version__ = "1.0.2"
+__author__ = "don_kanaille@posteo.de"
 
 # Create an instance of 'QApplication'
 crawler = QApplication(sys.argv)
